+++ src/hardware_layer/sensor_access/include/sensor_access/mock_imu_sensor.h
@@ -7,9 +7,9 @@
 class MockImuSensor : public ImuSensorInterface
 {
 public:
-  MOCK_METHOD0(getX, float(void));
-  MOCK_METHOD0(getY, float(void));
-  MOCK_METHOD0(getOmega, float(void));
+  MOCK_METHOD0(getX, double(void));
+  MOCK_METHOD0(getY, double(void));
+  MOCK_METHOD0(getOmega, double(void));
   MOCK_METHOD0(receiveData, ReadableSensors::ReadStatus());  // this is an "uninteresting function"
 };
